# src/main/resources/static/v3/api-docs/swagger.yaml
openapi: 3.1.1
info:
  title: Dataportal Backend REST API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 7.3.0
...
servers:
  - url: https://to.be.defined
    variables:
      basePath:
        default: /api/v5
...
paths:
  /query/data:
    post:
      tags:
        - dataquery
      summary: Store a dataquery in the backend.
      description: The query will only be stored, nothing will be dispatched. Does not have to pass any validation and may be incomplete. May or may not contain result.
      operationId: storeDataQuery
      requestBody:
        description: Dataquery
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Dataquery"
        required: true
      responses:
        201:
          description: Dataquery successfully stored
          headers:
            Location:
              description: Path to the result of your newly created dataquery
              schema:
                type: string
                format: url
        401:
          description: Unauthorized - please login first
        422:
          description: Invalid input
      security:
        - dataportal_auth:
            - user
    get:
      tags:
        - dataquery
      summary: Get the list of the calling users dataqueries
      description: This returns a list with basic information about the queries. Id, label (if present) and creation date. Basically for all parameters I'm not sure if we need them (at least right away)
      operationId: getDataQueryList
      parameters:
        - name: include-temporary
          in: query
          description: if set to true, queries spawned from other users are contained
          required: false
          schema:
            type: boolean
            default: false
        - name: skip-validation
          in: query
          description: If true, do not validate the query and do not include a list of invalid terms. TODO - maybe invert this?
          required: false
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataQueryListEntry"
        401:
          description: Unauthorized - please login first
      security:
        - dataportal_auth:
            - user
  /query/data/by-user/{userId}:
    post:
      tags:
        - dataquery
      summary: Spawn a query for another user
      operationId: createDataqueryForUser
      parameters:
        - name: ttl
          in: query
          description: How long until the query gets deleted (ISO 8601 Duration)
          required: true
          schema:
            type: string
      requestBody:
        description: Dataquery
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Dataquery"
        required: true
      responses:
        201:
          description: Dataquery successfully stored
          headers:
            Location:
              description: Path to the result of your newly created dataquery
              schema:
                type: string
                format: url
        401:
          description: Unauthorized - please login first
        403:
          description: Storage exceeded
      security:
        - dataportal_auth:
            - admin
    get:
      tags:
        - dataquery
      summary: Finds query summary (id, label, lastModified) of all queries of one user
      operationId: findDataQueriesByUser
      parameters:
        - name: userId
          in: path
          description: User to filter by (keycloak id)
          required: true
          schema:
            type: string
        - name: include-temporary
          in: query
          description: if set to true, queries spawned from other users are contained
          required: false
          schema:
            type: boolean
            default: false
        - name: skip-validation
          in: query
          description: If true, do not validate the query and do not include a list of invalid terms. TODO - maybe invert this?
          required: false
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DataQueryListEntry"
        401:
          description: Unauthorized - please login first
        403:
          description: Forbidden - insufficient access rights
        404:
          description: User not found
      security:
        - dataportal_auth:
            - admin
  /query/data/{queryId}:
    get:
      tags:
        - dataquery
      summary: Read dataquery by ID
      description: Returns a single dataquery.
      operationId: getDataQueryById
      parameters:
        - name: queryId
          in: path
          description: ID of query to return
          required: true
          schema:
            type: integer
            format: int64
        - name: skip-validation
          in: query
          description: If true, do not validate the query and do not include a list of invalid terms
          required: false
          schema:
            type: boolean
            default: false
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dataquery"
        401:
          description: Unauthorized - please login first
        403:
          description: Forbidden - insufficient access rights
        404:
          description: Query not found
      security:
        - dataportal_auth:
            - user
    put:
      tags:
        - dataquery
      summary: Update a dataquery in the backend.
      description: The query will be updated. It must be provided completely, there are no partial updates.
      operationId: updateDataQuery
      requestBody:
        description: CRTDL
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Dataquery"
        required: true
      responses:
        204:
          description: Dataquery successfully updated
          content: {}
        401:
          description: Unauthorized - please login first
        403:
          description: Forbidden - insufficient access rights
        404:
          description: The dataquery could not be found
      security:
        - dataportal_auth:
            - user
    delete:
      tags:
        - dataquery
      summary: Delete a dataquery
      operationId: deleteDataQuery
      parameters:
        - name: queryId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        204:
          description: No content
          content:
            {}
        401:
          description: Unauthorized - please login first
        403:
          description: Forbidden - insufficient access rights
        404:
          description: The dataquery could not be found
      security:
        - dataportal_auth:
            - user
...
  /query/data/convert/crtdl:
    post:
      tags:
        - dataquery
      summary: Convert a CRTDL to an zip file with csv contents.
      description: Nothing will be stored, only a zip file will be generated
      operationId: convertCrtdlToCsv
      requestBody:
        description: CRTDL
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CRTDL"
        required: true
      responses:
        200:
          description: CRTDL successfully converted
          headers:
            Content-Disposition:
              schema:
                type: string
                examples:
                  - attachment; filename=untitled.zip
            Content-Type:
              schema:
                type: string
                examples:
                  application/zip
        400:
          description: Bad request - validation of the CRTDL failed
        401:
          description: Unauthorized - please login first
        422:
          description: Invalid input
      security:
        - dataportal_auth:
            - user
...
  /query/data/query-slots:
    get:
      tags:
        - dataquery
      summary: Show how many data query slots a user already used and how many he has left.
      operationId: getDataQuerySlots
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DataQuerySlots"
      security:
        - dataportal_auth:
            - user
  /query/validate:
    post:
      tags:
        - dataquery
      summary: Validates a submitted CRTDL or just CCDL? to check for schema violations or invalid termCodes
      description: It could be useful to split this in ccdl and dataquery or at least offer a parameter to determine which should be checked.
      operationId: validateQuery
      requestBody:
        description: Query to validate
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CRTDL"
        required: true
      responses:
        200:
          description: Query adheres to json schema. If invalid termCodes are present, they will be in the response.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CRTDL"
        400:
          description: Query does not adhere to json schema
        401:
          description: Unauthorized - please login first
...
components:
  schemas:
    Dataquery:
      type: object
      required:
        - content
      properties:
        label:
          type: string
          examples:
            - my query
        comment:
          type: string
          examples:
            - this can be a longer text explaining what this query is for
        totalNumberOfResults:
          type: integer
        content:
          type: object
          $ref: "#/components/schemas/CRTDL"
    CRTDL:
      type: object
      properties:
        display:
          type: string
          examples:
            - Example CRTDL
        cohortDefinition:
          type: object
          $ref: "#/components/schemas/CCDL"
        dataExtraction:
          type: object
          $ref: "#/components/schemas/DataExtractionQuery"
    DataExtractionQuery:
      type: object
      properties:
        id:
          type: string
          examples:
            - 123
        attributeGroups:
          type: array
          items:
            type: string
            examples:
              - to be done correctly...for now just a placeholder because it doesn't matter for this conversation
    DataQueryListEntry:
      type: object
      required:
        - id
        - label
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        comment:
          type: string
        lastModified:
          type: string
          format: date-time
        resultSize:
          type: integer
        ccdl:
          type: object
          properties:
            exists:
              type: boolean
            isValid:
              type: boolean
        dataExtraction:
          type: object
          properties:
            exists:
              type: boolean
            isValid:
              type: boolean
...
    CCDL:
      type: object
      required:
        - version
        - inclusionCriteria
      properties:
        version:
          type: string
          format: uri
          description: The json schema version
          examples:
            - http://to_be_decided.com/draft-1/schema#
        display:
          type: string
          examples:
            - foobar
        inclusionCriteria:
          type: array
          items:
            $ref: "#/components/schemas/CriterionList"
        exclusionCriteria:
          type: array
          items:
            $ref: "#/components/schemas/CriterionList"
    DataQuerySlots:
      type: object
      required:
        - used
        - total
      properties:
        used:
          type: integer
          description: The amount of used dataquery slots for a user.
          examples:
            - 2
        total:
          type: integer
          description: The total amount of dataquery slots per user.
          examples:
            - 10
    TermCode:
      description: The termCode defines a concept based on a coding system (i.e. LOINC). The triplet of code, system and version identify the concept.
      type: object
      required:
        - code
        - system
        - display
      properties:
        code:
          type: string
          examples:
            - 119373006
        system:
          type: string
          examples:
            - http://snomed.info/sct
        version:
          type: string
          examples:
            - http://snomed.info/sct/900000000000207008/version/20210731
        display:
          type: string
          examples:
            - Amniotic fluid specimen (specimen)
    beforeDate:
      type: string
      format: date-time
    afterDate:
      type: string
      format: date-time
    TimeRestriction:
      anyOf:
        - $ref: "#/components/schemas/beforeDate"
        - $ref: "#/components/schemas/afterDate"
    Unit:
      type: object
      required:
        - code
        - display
      properties:
        code:
          type: string
        display:
          type: string
    AttributeFilter:
      type: object
      description: An AttributeFilter requires different properties, depending on the type. Please refer to the JSON Schema for this.
      required:
        - type
      properties:
        attributeCode:
          $ref: "#/components/schemas/TermCode"
    ValueFilter:
      type: object
      description: A ValueFilter requires different properties, depending on the type. Please refer to the JSON Schema for this.
      required:
        - type
      properties:
        type:
          type: string
          enum:
            - concept
            - quantity-comparator
            - quantity-range
        selectedConcepts:
          type: array
          items:
            $ref: "#/components/schemas/TermCode"
        comparator:
          type: string
          enum:
            - eq
            - ue
            - le
            - lt
            - ge
            - gt
        unit:
          $ref: "#/components/schemas/Unit"
        value:
          type: number
          format: double
        minValue:
          type: number
          format: double
        maxValue:
          type: number
          format: double
    Criterion:
      type: object
      required:
        - termCodes
      properties:
        termCodes:
          type: array
          items:
            $ref: "#/components/schemas/TermCode"
        attributeFilters:
          type: array
          items:
            $ref: "#/components/schemas/AttributeFilter"
        valueFilter:
          $ref: "#/components/schemas/ValueFilter"
        timeRestriction:
          $ref: "#/components/schemas/TimeRestriction"
        issues:
          type: array
          items:
            $ref: "#/components/schemas/ValidationIssue"
    CriterionList:
      type: array
      items:
        $ref: "#/components/schemas/Criterion"
...
    ValidationIssue:
      type: object
      required:
        - code
        - detail
      properties:
        code:
          type: string
          examples:
            - VAL-20001
        detail:
          type: string
          examples:
            - The combination of context and termcode(s) is not found.
...
  securitySchemes:
    dataportal_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: http://to.be.defined/auth
          scopes:
            user: Dataportal user role
            admin: Dataportal admin role